Extract ShortFeedback from PlatformFooter
